+++ CHANGELOG.md
@@ -1,4 +1,24 @@
 # Changelog
+## [0.3.1](https://github.com/aacn/eslint-plugin-tailwind-classname-order/releases/tag/v0.3.0) - 2023-01-19
+### Updated
+- Adjusted README
+
+## [0.3.0](https://github.com/aacn/eslint-plugin-tailwind-classname-order/releases/tag/v0.3.0) - 2023-01-19
+### Updated
+- Removed unecessary files from lib
+- Changed package name to **@aacn.eu/eslint-plugin-tailwind-classname-order**
+### Added
+- Added support for [tailwind-rn](https://www.npmjs.com/package/tailwind-rn)
+- Added support for expressions
+```js
+ // tailwind-rn
+<div style={tailwind("px-3 mt-4")}></div>
+
+// expressions
+<div className={"px-3 mt-4"}></div>
+<div className={`px-3 mt-4 ${myVar}`}></div>
+```
+
 ## [0.2.7](https://github.com/aacn/eslint-plugin-tailwind-classname-order/releases/tag/v0.2.7) - 2022-09-08
 ### Fixed
 - Fixed a bug, where the plugin couldn't interpret classNames that had the !important modifier added to it (e.g. !text-green).